+++ harmonia_element/tests.lua
@@ -1,0 +1,6 @@
+local mod = assert(harmonia_element)
+
+--
+--
+--
+mod:require("tests/element_system_test.lua")