+++ harmonia_element/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.3.0
+
+* Added crafting events:
+  * `craft.started`
+  * `craft.error`
+  * `craft.completed`
+
 # 0.2.0
 
 * Added basic element crafting system